converterなどの削除
