+++ application/Makefile
@@ -6,6 +6,7 @@ GITHUB_SHA ?= local
 
 .PHONY: test
 test:
+		go clean -testcache
 		go test ./...
