+++ application/api/service/output/messate.go
@@ -1,0 +1,5 @@
+package output
+
+type Message struct {
+	Message string
+}
