+++ application/api/service/output/user.go
@@ -1,0 +1,16 @@
+package output
+
+type UserInfoResponse struct {
+	UserID          string
+	AuthToken       string
+	Platform        string
+	PremiumPlan     bool
+	IOSPlatformInfo IOSPlatformInfoResponse
+}
+
+type IOSPlatformInfoResponse struct {
+	PushToken                string
+	PushTokenSNSEndpoint     string
+	VoIPPushToken            string
+	VoIPPushTokenSNSEndpoint string
+}
